Update popup when user tries to click on text editor
