+++ webapp/src/notification.tsx
@@ -3,6 +3,7 @@
 import * as React from "react";
 import * as data from "./data";
 import * as sui from "./sui";
+import { jsxLF } from "./util";
 
 import Cloud = pxt.Cloud;
 
@@ -126,9 +127,9 @@ export class NotificationBanner extends data.Component<ISettingsProps, {}> {
         const showExperiments = pxt.editor.experiments.someEnabled() && !/experiments=1/.test(window.location.href);
         const showWinAppBanner = pxt.appTarget.appTheme.showWinAppDeprBanner && pxt.BrowserUtils.isWinRT();
 
-        const errMsg = this.jsxLF(
+        const errMsg = jsxLF(
             "This app is no longer supported. For the latest updates, {0} to install our new app! {1}",
-            <sui.Link className="link" ariaLabel={lf("click here")} onClick={this.getNewAppClick}>{lf("click here")}</sui.Link>,            
+            <sui.Link className="link" ariaLabel={lf("click here")} onClick={this.getNewAppClick}>{lf("click here")}</sui.Link>,
             <sui.Link className="link" ariaLabel={lf("More info")} onClick={this.handleBannerClick}>{lf("Learn More")}</sui.Link>
         );
 
@@ -162,32 +163,5 @@ export class NotificationBanner extends data.Component<ISettingsProps, {}> {
         }
 
         return <div></div>;
-    }
-
-    // Based on https://github.com/microsoft/pxt/blob/master/react-common/components/util.tsx#L15.
-    jsxLF(loc: string, ...rest: JSX.Element[]) {
-        const indices: number[] = [];
-
-        loc.replace(/\{\d\}/g, match => {
-            indices.push(parseInt(match.substr(1, 1)));
-            return match;
-        });
-
-        const out: JSX.Element[] = [];
-
-        let parts: string[];
-
-        let i = 0;
-
-        for (const index of indices) {
-            parts = loc.split(`{${index}}`);
-            pxt.U.assert(parts.length === 2);
-            out.push(<span key={i++}>{parts[0]}</span>);
-            out.push(<span key={i++}>{rest[index]}</span>);
-            loc = parts[1];
-        }
-        out.push(<span key={i++}>{loc}</span>);
-
-        return out;
     }
 }
