+++ webapp/src/util.tsx
@@ -1,0 +1,28 @@
+import React = require("react");
+
+// Based on https://github.com/microsoft/pxt/blob/master/react-common/components/util.tsx#L15.
+export function jsxLF(loc: string, ...rest: JSX.Element[]) {
+    const indices: number[] = [];
+
+    loc.replace(/\{\d\}/g, match => {
+        indices.push(parseInt(match.substr(1, 1)));
+        return match;
+    });
+
+    const out: JSX.Element[] = [];
+
+    let parts: string[];
+
+    let i = 0;
+
+    for (const index of indices) {
+        parts = loc.split(`{${index}}`);
+        pxt.U.assert(parts.length === 2);
+        out.push(<span key={i++}>{parts[0]}</span>);
+        out.push(<span key={i++}>{rest[index]}</span>);
+        loc = parts[1];
+    }
+    out.push(<span key={i++}>{loc}</span>);
+
+    return out;
+}
